+++ src/bittytax/conv/parsers/kucoin.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # (c) Nano Nano Ltd 2020
 
+import re
 from decimal import Decimal
 from typing import TYPE_CHECKING
 
@@ -212,12 +213,15 @@ def parse_kucoin_withdrawals(
 
 
 def parse_kucoin_deposits_withdrawals_v2(
-    data_row: "DataRow", _parser: DataParser, **_kwargs: Unpack[ParserArgs]
+    data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[ParserArgs]
 ) -> None:
     row_dict = data_row.row_dict
-    data_row.timestamp = DataParser.parse_timestamp(
-        row_dict["Time(UTC+08:00)"], tz="Asia/Singapore"
-    )
+
+    if parser.args:
+        timestamp_hdr = parser.args[0].group(1)
+        utc_offset = parser.args[0].group(2)
+
+    data_row.timestamp = DataParser.parse_timestamp(f"{row_dict[timestamp_hdr]} {utc_offset}")
 
     if row_dict["Status"] != "SUCCESS":
         return
@@ -230,7 +234,7 @@ def parse_kucoin_deposits_withdrawals_v2(
             sell_asset=row_dict["Coin"],
             fee_quantity=Decimal(row_dict["Fee"]),
             fee_asset=row_dict["Coin"],
-            wallet=row_dict["Withdrawal Address/Account"],
+            wallet=WALLET,
         )
     else:
         data_row.t_record = TransactionOutRecord(
@@ -245,12 +249,15 @@ def parse_kucoin_deposits_withdrawals_v2(
 
 
 def parse_kucoin_staking_income(
-    data_row: "DataRow", _parser: DataParser, **_kwargs: Unpack[ParserArgs]
+    data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[ParserArgs]
 ) -> None:
     row_dict = data_row.row_dict
-    data_row.timestamp = DataParser.parse_timestamp(
-        row_dict["Time(UTC+08:00)"], tz="Asia/Singapore"
-    )
+
+    if parser.args:
+        timestamp_hdr = parser.args[0].group(1)
+        utc_offset = parser.args[0].group(2)
+
+    data_row.timestamp = DataParser.parse_timestamp(f"{row_dict[timestamp_hdr]} {utc_offset}")
 
     data_row.t_record = TransactionOutRecord(
         TrType.STAKING,
@@ -376,7 +383,7 @@ DataParser(
     [
         "UID",
         "Account Type",
-        "Time(UTC+08:00)",
+        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
         "Remarks",
         "Status",
         "Fee",
@@ -394,7 +401,7 @@ DataParser(
     [
         "UID",
         "Account Type",
-        "Time(UTC+08:00)",
+        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
         "Remarks",
         "Status",
         "Fee",
@@ -414,7 +421,7 @@ DataParser(
         "UID",
         "Account Type",
         "Order ID",
-        "Time(UTC+08:00)",
+        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
         "Staked Coin",
         "Product Type",
         "Product Name",